+++ sdk/appcenter-crashes/src/androidTest/java/com/microsoft/appcenter/crashes/ingestion/models/ErrorModelTest.java
@@ -72,6 +72,7 @@ public class ErrorModelTest {
             exception3.setStackTrace(exception1.getStackTrace());
             exception3.setFrames(exception1.getFrames());
             exception3.setWrapperSdkName(exception1.getWrapperSdkName());
+            exception3.setMinidumpFilePath(exception1.getMinidumpFilePath());
             errorLog2.setException(exception3);
             checkEquals(errorLog1, errorLog2);
 
@@ -309,6 +310,17 @@ public class ErrorModelTest {
                 checkNotEquals(exception1, exception2);
 
                 exception2.setStackTrace(exception1.getStackTrace());
+                checkEquals(exception1, exception2);
+            }
+            {
+                exception1.setMinidumpFilePath("1");
+                checkNotEquals(exception1, exception2);
+                checkExceptions(serializer, errorLog1, errorLog2, exception1, exception2);
+
+                exception2.setMinidumpFilePath("2");
+                checkNotEquals(exception1, exception2);
+
+                exception2.setMinidumpFilePath(exception1.getMinidumpFilePath());
                 checkEquals(exception1, exception2);
             }
             {